HOL-40 Clean-up and rebase fix.
