+++ unity/Holo/Assets/Editor/ModelLoader.cs
@@ -4,25 +4,29 @@ using UnityEngine;
 
 class ModelLoader
 {
+    /* Loads a model in batchmode or multiple models in Editor and converts them into an AssetBundle.
+     * To use in batchmode: "<Path to Unity.exe>" -quit -batchmode -logFile "<Path to the logfile>"  
+    * -executeMethod ModelLoader.LoadModel -rootDirectory "<Directory of the folder which stores the meshes>" 
+    */
     [MenuItem("Holo/Convert model to a AssetBundle's GameObject")]
-    public static void PreprocessModel()
+    public static void LoadModel()
     {
         SingleModel importedModel = new SingleModel();
         AssetBundleCreator assetBundleCreator = new AssetBundleCreator();
-        string rootDirectory = "";
-        bool loadModel = true;
 
+        bool loadModel = true;
         while (loadModel)
         {
-            rootDirectory = EditorUtility.OpenFolderPanel("Select model's root folder", Application.dataPath, "");
-            if (String.IsNullOrEmpty(rootDirectory))
-            {
-                throw new ArgumentException("Path cannot be null!");
-            }
-
-            importedModel.GetModelData(rootDirectory);
+            importedModel.GetModelData();
             assetBundleCreator.Create(importedModel);
-            loadModel = EditorUtility.DisplayDialog("", "Do you want to load another model?", "Yes", "No");
+            if (Application.isBatchMode)
+            {
+                loadModel = true;
+            }
+            else
+            {
+                loadModel = EditorUtility.DisplayDialog("", "Do you want to load another model?", "Yes", "No");
+            }
         }
     }
 }
